Debugged ImageJ integration.  Now possible to run a method from a parameter file from withing ImageJ as well as to edit the parameter files graphically.  Images are now internally ImageJ ImagePlus for pixel data.  Added the LocalMaximumSeparabilityThresholdingFilter, which greatly improves robustness of centromere finding to background noise.  Resolves issue 7.

+++ src/edu/stanford/cfuller/imageanalysistools/filter/ConvolutionFilter.java
@@ -189,7 +189,7 @@ public class ConvolutionFilter extends Filter {
         oldMin  = 0;
 
         for (ImageCoordinate ic : orig) {
-            orig.setValue(ic, (rowImage[ic.get(ImageCoordinate.Y)][ic.get(ImageCoordinate.X)] - newMin)*scaleFactor + oldMin);
+            orig.setValue(ic, (float) ((rowImage[ic.get(ImageCoordinate.Y)][ic.get(ImageCoordinate.X)] - newMin)*scaleFactor + oldMin));
         }
         
         orig.clearBoxOfInterest();
